Simplify logging about device availability
I'm still not sure if and how I will use this data. If
the timeouts prove good enough for most casees I may just leave them
rather than having to coordinate information across threads/mail
loop iterations. But, if that doesn't work, I'm going to want
to have an understanding of what the system is doing with my
UBW board