Structure pysprinklers as an installable module
This includes moving everything into pysprinklers,
the module subdir, and adding a simple setup.py
file. This was necessary to make imports work
when I split out the platform.

Incidentally, this splits out the platform. The
platform is the module in charge of basic functionality
like talking to the DB or UBW. Since this is
very common between the webserver and glib, it makes
sense to factor it out